# internal/models/user.go
package models

import "go.mongodb.org/mongo-driver/bson/primitive"

// Usuario representa un usuario almacenado en MongoDB
type Usuario struct {
	ID       primitive.ObjectID `bson:"_id,omitempty" json:"id"`
	Nombre   string             `bson:"nombre" json:"nombre"`
	Email    string             `bson:"email" json:"email"`
	Password string             `bson:"password" json:"password,omitempty"`
}